Do some major refactoring
