+++ backend/cmd/cluster_counter/main.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"encoding/json"
 	count "github.com/Avi18971911/Augur/pkg/count/service"
+	augurElasticsearch "github.com/Avi18971911/Augur/pkg/elasticsearch/bootstrapper"
 	"github.com/Avi18971911/Augur/pkg/elasticsearch/client"
-	augurElasticsearch "github.com/Avi18971911/Augur/pkg/elasticsearch/db_model"
 	"github.com/elastic/go-elasticsearch/v8"
 	"go.uber.org/zap"
 	_ "google.golang.org/grpc/encoding/gzip"
