+++ backend/pkg/elasticsearch/client/augur_client.go
@@ -1,19 +1,8 @@
 package client
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
-	"fmt"
-	"io"
-	"log"
-	"strings"
-	"time"
-
-	"github.com/Avi18971911/Augur/pkg/elasticsearch/model"
-	logModel "github.com/Avi18971911/Augur/pkg/log/model"
 	"github.com/elastic/go-elasticsearch/v8"
-	"github.com/elastic/go-elasticsearch/v8/esapi"
 )
 
 const SearchResultSize = 10
@@ -61,421 +50,4 @@ type AugurClientImpl struct {
 
 func NewAugurClientImpl(es *elasticsearch.Client, refreshRate RefreshRate) *AugurClientImpl {
 	return &AugurClientImpl{es: es, refreshRate: string(refreshRate)}
-}
-
-// TODO: Make functions blocking to avoid race conditions
-func (a *AugurClientImpl) BulkUpdate(
-	ctx context.Context,
-	ids []string,
-	fieldList []map[string]interface{},
-	index string,
-) error {
-	var buf bytes.Buffer
-	for i, fields := range fieldList {
-		update := map[string]interface{}{
-			"update": map[string]interface{}{
-				"_id": ids[i],
-			},
-		}
-		metaJSON, err := json.Marshal(update)
-		if err != nil {
-			return fmt.Errorf("error marshaling update: %w", err)
-		}
-		buf.Write(metaJSON)
-		buf.WriteByte('\n')
-
-		fieldJSON, err := json.Marshal(map[string]interface{}{"doc": fields})
-		if err != nil {
-			return fmt.Errorf("error marshaling field to update: %w", err)
-		}
-		buf.Write(fieldJSON)
-		buf.WriteByte('\n')
-	}
-
-	res, err := a.es.Bulk(
-		bytes.NewReader(buf.Bytes()),
-		a.es.Bulk.WithIndex(index),
-		a.es.Bulk.WithContext(ctx),
-		a.es.Bulk.WithRefresh(a.refreshRate),
-	)
-	if err != nil {
-		return fmt.Errorf("failed to update in Elasticsearch: %w", err)
-	}
-	defer res.Body.Close()
-	if res.IsError() {
-		return fmt.Errorf("bulk update error: %s", res.String())
-	}
-	return nil
-}
-
-func (a *AugurClientImpl) Upsert(
-	ctx context.Context,
-	upsertScript map[string]interface{},
-	index string,
-	id string,
-) error {
-	var buf bytes.Buffer
-	upsertJSON, err := json.Marshal(upsertScript)
-	if err != nil {
-		return fmt.Errorf("error marshaling upsert: %w", err)
-	}
-	buf.Write(upsertJSON)
-
-	res, err := a.es.Update(
-		index, id,
-		bytes.NewReader(buf.Bytes()),
-		a.es.Update.WithContext(ctx),
-		a.es.Update.WithRefresh(a.refreshRate),
-	)
-	if err != nil {
-		return fmt.Errorf("failed to upsert in Elasticsearch: %w", err)
-	}
-	defer res.Body.Close()
-	if res.IsError() {
-		return fmt.Errorf("upsert error: %s", res.String())
-	}
-	return nil
-}
-
-func (a *AugurClientImpl) BulkIndex(
-	ctx context.Context,
-	data []map[string]interface{},
-	metaInfo []map[string]interface{},
-	index string,
-) error {
-	var buf bytes.Buffer
-	for i, d := range data {
-		var meta map[string]interface{}
-		if metaInfo != nil && i < len(metaInfo) {
-			meta = metaInfo[i]
-		} else {
-			// empty meta for bulk index
-			meta = map[string]interface{}{"index": map[string]interface{}{}}
-		}
-		metaJSON, err := json.Marshal(meta)
-		if err != nil {
-			return fmt.Errorf("error marshaling meta to bulk index: %w", err)
-		}
-		buf.Write(metaJSON)
-		buf.WriteByte('\n')
-
-		dataJSON, err := json.Marshal(d)
-		if err != nil {
-			return fmt.Errorf("error marshaling data to bulk index: %w", err)
-		}
-		buf.Write(dataJSON)
-		buf.WriteByte('\n')
-	}
-	var res *esapi.Response
-	var err error
-	if len(index) > 0 {
-		res, err = a.es.Bulk(
-			bytes.NewReader(buf.Bytes()),
-			a.es.Bulk.WithIndex(index),
-			a.es.Bulk.WithContext(ctx),
-			a.es.Bulk.WithRefresh(a.refreshRate),
-		)
-	} else {
-		res, err = a.es.Bulk(
-			bytes.NewReader(buf.Bytes()),
-			a.es.Bulk.WithContext(ctx),
-			a.es.Bulk.WithRefresh(a.refreshRate),
-		)
-	}
-	if err != nil {
-		return fmt.Errorf("error bulk indexing: %w", err)
-	}
-	defer res.Body.Close()
-	if res.IsError() {
-		return fmt.Errorf("bulk index error: %s", res.String())
-	}
-	return nil
-}
-
-func (a *AugurClientImpl) Index(
-	ctx context.Context,
-	data map[string]interface{},
-	metaInfo map[string]interface{},
-	index string,
-) error {
-	if metaInfo == nil {
-		return a.BulkIndex(ctx, []map[string]interface{}{data}, nil, index)
-	}
-	return a.BulkIndex(ctx, []map[string]interface{}{data}, []map[string]interface{}{metaInfo}, index)
-}
-
-func (a *AugurClientImpl) Search(
-	ctx context.Context,
-	query string,
-	indices []string,
-	queryResultSize *int,
-) ([]map[string]interface{}, error) {
-	res, err := a.es.Search(
-		a.es.Search.WithContext(ctx),
-		a.es.Search.WithIndex(indices...),
-		a.es.Search.WithBody(strings.NewReader(query)),
-		a.es.Search.WithPretty(),
-		a.es.Search.WithSize(getQuerySize(queryResultSize)),
-	)
-
-	if err != nil {
-		return nil, fmt.Errorf("failed to execute query: %w", err)
-	}
-	defer res.Body.Close()
-
-	if res.IsError() {
-		return nil, fmt.Errorf("failed to execute query: %s", res.String())
-	}
-
-	var esResponse model.EsResponse
-	if err := json.NewDecoder(res.Body).Decode(&esResponse); err != nil {
-		return nil, fmt.Errorf("failed to decode response body: %w", err)
-	}
-
-	var results []map[string]interface{}
-	for _, hit := range esResponse.Hits.HitArray {
-		results = append(results, hit.Source)
-		results[len(results)-1]["_id"] = hit.ID
-	}
-
-	return results, nil
-}
-
-func (a *AugurClientImpl) Count(
-	ctx context.Context,
-	query string,
-	indices []string,
-) (int64, error) {
-	res, err := a.es.Count(
-		a.es.Count.WithContext(ctx),
-		a.es.Count.WithIndex(indices...),
-		a.es.Count.WithBody(strings.NewReader(query)),
-		a.es.Count.WithPretty(),
-	)
-
-	if err != nil {
-		return 0, fmt.Errorf("failed to execute query: %w", err)
-	}
-	defer res.Body.Close()
-
-	if res.IsError() {
-		return 0, fmt.Errorf("failed to execute query: %s", res.String())
-	}
-
-	var countResponse model.CountResponse
-	if err := json.NewDecoder(res.Body).Decode(&countResponse); err != nil {
-		return 0, fmt.Errorf("failed to decode response body: %w", err)
-	}
-
-	return int64(countResponse.Count), nil
-}
-
-func (a *AugurClientImpl) UpdateByQuery(
-	ctx context.Context,
-	query string,
-	indices []string,
-) error {
-	res, err := a.es.UpdateByQuery(
-		indices,
-		a.es.UpdateByQuery.WithBody(strings.NewReader(query)),
-		a.es.UpdateByQuery.WithContext(ctx),
-		a.es.UpdateByQuery.WithRefresh(a.refreshRate == string(Immediate)),
-	)
-	if err != nil {
-		return fmt.Errorf("failed to update by query in Elasticsearch: %w", err)
-	}
-	defer res.Body.Close()
-	if res.IsError() {
-		return fmt.Errorf("update by query error: %s", res.String())
-	}
-	return nil
-}
-
-func (a *AugurClientImpl) SearchAfter(
-	ctx context.Context,
-	query map[string]interface{},
-	indices []string,
-	orderingParams []map[string]string,
-	querySize *int,
-) (string, bool, error) {
-	keepAlive := "1m"
-	pitRes, err := a.es.OpenPointInTime(
-		indices,
-		keepAlive,
-		a.es.OpenPointInTime.WithContext(ctx),
-	)
-	if err != nil {
-		return "", false, fmt.Errorf("failed to open point in time: %w", err)
-	}
-	defer pitRes.Body.Close()
-	if pitRes.IsError() {
-		return "", false, fmt.Errorf("failed to open point in time: %s", pitRes.String())
-	}
-	pitId, err := readPitBody(pitRes.Body)
-	if err != nil {
-		return "", false, fmt.Errorf("failed to read pit id: %w", err)
-	}
-
-	pitQuery := buildSearchWithPitQuery(query, pitId, orderingParams)
-	pitQueryJson, err := json.Marshal(pitQuery)
-	if err != nil {
-		return "", false, fmt.Errorf("failed to marshal pit query: %w", err)
-	}
-
-	res, err := a.es.Search(
-		a.es.Search.WithContext(ctx),
-		a.es.Search.WithIndex(indices...),
-		a.es.Search.WithBody(strings.NewReader(string(pitQueryJson))),
-		a.es.Search.WithSize(getQuerySize(querySize)),
-	)
-	if err != nil {
-		return "", false, fmt.Errorf("failed to execute query: %w", err)
-	}
-	defer res.Body.Close()
-	if res.IsError() {
-		return "", false, fmt.Errorf("failed to execute query: %s", res.String())
-	}
-}
-
-func (a *AugurClientImpl) closePointInTime(pitID string, ctx context.Context, err error) error {
-	var buf bytes.Buffer
-	body := map[string]interface{}{"pit_id": pitID}
-	if err := json.NewEncoder(&buf).Encode(body); err != nil {
-		log.Fatalf("Error encoding close PIT request body: %v", err)
-	}
-
-	res, err := a.es.ClosePointInTime(
-		a.es.ClosePointInTime.WithBody(&buf),
-		a.es.ClosePointInTime.WithContext(ctx),
-	)
-	if err != nil {
-		return fmt.Errorf("failed to close PIT: %w", err)
-	}
-	defer res.Body.Close()
-	if res.IsError() {
-		return fmt.Errorf("failed to close PIT: %s", res.String())
-	}
-	return nil
-}
-
-func ToMetaAndDataMap[T any](values []T) ([]map[string]interface{}, []map[string]interface{}, error) {
-	dataMap := make([]map[string]interface{}, len(values))
-	metaMap := make([]map[string]interface{}, len(values))
-	for i, v := range values {
-		data, err := json.Marshal(v)
-		if err != nil {
-			return nil, nil, fmt.Errorf("failed to marshal value to JSON: %w", err)
-		}
-		var mapStruct map[string]interface{}
-		if err := json.Unmarshal(data, &mapStruct); err != nil {
-			return nil, nil, fmt.Errorf("failed to unmarshal JSON to map: %w", err)
-		}
-
-		if id, ok := mapStruct["_id"]; ok {
-			delete(mapStruct, "_id")
-			metaMap[i] = map[string]interface{}{"index": map[string]interface{}{"_id": id}}
-		} else {
-			metaMap[i] = map[string]interface{}{"index": map[string]interface{}{}}
-		}
-		dataMap[i] = mapStruct
-	}
-	return metaMap, dataMap, nil
-}
-
-func NormalizeTimestampToNanoseconds(timestamp string) (time.Time, error) {
-	isUTC := strings.HasSuffix(timestamp, "Z")
-	if isUTC {
-		timestamp = strings.TrimSuffix(timestamp, "Z")
-	}
-
-	parts := strings.SplitN(timestamp, ".", 2)
-	if len(parts) == 2 {
-		fractionalPart := parts[1]
-
-		// 9 digits (nanosecond)
-		if len(fractionalPart) > 9 {
-			fractionalPart = fractionalPart[:9]
-		} else if len(fractionalPart) < 9 {
-			fractionalPart = fractionalPart + strings.Repeat("0", 9-len(fractionalPart))
-		}
-
-		timestamp = parts[0] + "." + fractionalPart
-	}
-
-	if isUTC {
-		timestamp += "Z"
-	}
-
-	layout := "2006-01-02T15:04:05.000000000Z"
-	return time.Parse(layout, timestamp)
-}
-
-// TODO: avoid this cumbersome function by using an elasticsearch client closer to logs
-func ConvertToLogDocuments(data []map[string]interface{}) ([]logModel.LogEntry, error) {
-	var docs []logModel.LogEntry
-
-	for _, item := range data {
-		doc := logModel.LogEntry{}
-
-		timestamp, ok := item["timestamp"].(string)
-		if !ok {
-			return nil, fmt.Errorf("failed to convert timestamp to string %v", item["timestamp"])
-		}
-
-		timestampParsed, err := NormalizeTimestampToNanoseconds(timestamp)
-		if err != nil {
-			return nil, fmt.Errorf("failed to convert timestamp '%s' to time.Time: %v", timestamp, err)
-		}
-
-		doc.Timestamp = timestampParsed
-
-		severity, ok := item["severity"].(string)
-		if !ok {
-			return nil, fmt.Errorf("failed to convert severity to string")
-		}
-
-		doc.Severity = logModel.Level(severity)
-
-		message, ok := item["message"].(string)
-		if !ok {
-			return nil, fmt.Errorf("failed to convert message to string")
-		}
-		doc.Message = message
-
-		service, ok := item["service"].(string)
-		if !ok {
-			return nil, fmt.Errorf("failed to convert service to string")
-		}
-		doc.Service = service
-
-		clusterId, ok := item["cluster_id"].(string)
-		if ok {
-			doc.ClusterId = clusterId
-		}
-
-		doc.Id = item["_id"].(string)
-		docs = append(docs, doc)
-	}
-
-	return docs, nil
-}
-
-func readPitBody(body io.ReadCloser) (string, error) {
-	var pitResponse map[string]any
-	if err := json.NewDecoder(body).Decode(&pitResponse); err != nil {
-		return "", fmt.Errorf("failed to decode pit response: %w", err)
-	}
-	pitId, ok := pitResponse["id"].(string)
-	if !ok {
-		return "", fmt.Errorf("failed to read pit id")
-	}
-	return pitId, nil
-}
-
-func getQuerySize(querySize *int) int {
-	if querySize == nil {
-		return SearchResultSize
-	} else {
-		return *querySize
-	}
 }
