+++ backend/pkg/trace/service/span_cluster_service.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	augurElasticsearch "github.com/Avi18971911/Augur/pkg/elasticsearch/bootstrapper"
 	"github.com/Avi18971911/Augur/pkg/elasticsearch/client"
-	augurElasticsearch "github.com/Avi18971911/Augur/pkg/elasticsearch/db_model"
 	"github.com/Avi18971911/Augur/pkg/trace/model"
 	"github.com/google/uuid"
 	"go.uber.org/zap"
